Make `tag` parameter optional on form rows
A row with a nil tag will use the name for a tag. This has been the
intended behavior all along. The implementations already used a nil
coalescing operator to that effect. But the input parameter on
constructing a form row was neither Optional, nor had a default `nil`
value.